# Imporved Painting, LineDrawing Performance