# Summery
i founded inefficient logic when i'm reviewing color clustering process. color clustering and line drawing process.

so i imporved color clustring process about 99%, and line drawing process about 70%.

## What I Work
* **Improved Color Clustring Performance.**
     i replaced clustring color logic with built-in function of opencv.
* **improved Line Drawing Performance.**
* **Removed Duplicated Logics.**

## Result
i did test 10 times using Time module to figure out how many time improved. the result is shown below.

### Color Clustring Performance
NO | BEFORE(s) | AFTER(s)
 -- | -- | --
1 | 1.88826 | 0.00898
2 | 1.88634 | 0.00798
3 | 1.90329 | 0.00997
4 | 1.88940 | 0.00798
5 | 1.88760 | 0.00898
6 | 1.87907 | 0.00997
7 | 1.87896 | 0.00898
8 | 1.91196 | 0.00798
9 | 1.89863 | 0.00898
10 | 1.87554 | 0.00898

In this result, we can see the average time reduced by 99%. (1.88s -> 0.008s)

### Line Drawing Performance
NO | BEFORE(s) | AFTER(s)
-- | -- | --
1 | 0.25919 | 0.08078
2 | 0.25919 | 0.08078
3 | 0.26628 | 0.08477
4 | 0.25972 | 0.07979
5 | 0.27278 | 0.07834

In this result, we can see the average time reduced by 70%. (0.26343s -> 0.08089s)

**BY THE WAY**, By replacing it with an index map from photos, I think the intent of the class has changed significantly. This is because in order to get the index map, painting process need to perform beforehand. Let me know what you think about it.